Removed resultsdetail and renamed results folders and filenames
