--- client/components/ResultDetails/ResultDetails.jsx
@@ -1,16 +1,0 @@
-import React, { PropTypes } from 'react';
-
-const ResultDetails = (props) => {
-  return (
-    <div className="container">
-      <p>
-        Results Details Page
-      </p>
-    </div>
-  );
-};
-
-ResultDetails.propTypes = {
-};
-
-export default ResultDetails;
