+++ client/entry.js
@@ -9,7 +9,7 @@ import HomeContainer from './components/Home/HomeContainer.jsx';
 import MainContainer from './components/Main/MainContainer.jsx';
 import NewScenarioContainer from './components/NewScenario/NewScenarioContainer.jsx';
 
-import LiveResultsContainer from './components/LiveResults/LiveResultsContainer.jsx';
+import ResultsContainer from './components/Results/ResultsContainer.jsx';
 
 const router = (
   <Provider store={store}>
@@ -18,7 +18,7 @@ const router = (
         <IndexRoute component={HomeContainer} />
         <Route path="/main" component={MainContainer} />
         <Route path="/new-scenario" component={NewScenarioContainer} />
-        <Route path="/live-results" component={LiveResultsContainer} />
+        <Route path="/results" component={ResultsContainer} />
       </Route>
     </Router>
   </Provider>
