+++ client/redux/actionCreators/results-helpers.js
